+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Properly notify UI of unit removal from inventory
 - Stopped Legonis Machina boss overriding certain difficulty settings
 - Queller not always loading the right personality when playing as Legonis Machina
+- Remove pointless Commander modifiers from Cluster systems
 
 ## v6.2.0 - 2026-07-13